JVM Warmup to better the class loading
A simple way to use the static block to load the class rather than relying on java invoking it through 'main'. Of course, in 'static' block, the class has been loaded already and hence accessing it's method is way more faster. Look for comments in the 'JVMWarmupDemo.java' for more information.